Moved filter to a new file.
